requirements.txt
crewai
python-dotenv
langchain_google_genai
smtplib
email
google-auth
google-generativeai
